develop --> master for release 1.1.0

+++ src/main/java/org/rsultan/core/clustering/medoidshift/MedianShift.java
@@ -1,13 +1,12 @@
 package org.rsultan.core.clustering.medoidshift;
 
-import static org.rsultan.core.clustering.type.MedoidType.MEAN;
 import static org.rsultan.core.clustering.type.MedoidType.MEDIAN;
 
 import org.rsultan.dataframe.Dataframe;
 
 public class MedianShift extends MedoidShift {
 
-  public MedianShift(long bandwidth, long epoch) {
+  public MedianShift(double bandwidth, long epoch) {
     super(bandwidth, epoch, MEDIAN);
   }
